> benchmark for coin store runs faster when database is located on SSD (129.9s vs 194.3s)
> benchmark for coin store also runs faster when database is located on SPINDLE (2135.5s vs 2546.2s)
> the coin store benchmark application was slightly improved to allow SQL logging, tracking total exec time and seeding the random
> using a partial index on block_records(is_peak) allows for a much smaller index -> less IO

This is achieved by
> using memory for temp tables
> using bulk insert from temp table instead of single rows into the coin store
> using a single update statement to change the spent index and flag of many coins at the same time


`#SSD original code
Building database
78.3017s, MOSTLY ADDITIONS additions: 400400 removals: 20000
9.2151s, MOSTLY REMOVALS additions: 600 removals: 140000
85.1686s, FULLBLOCKS additions: 400400 removals: 400000
1.4034s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
1.6274s, GET RECORDS BY NAMES without spent 200 lookups found 20669 coins in total
18.5551s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 194.2713s
`

`
#SSD bulk insert and mass update
Building database
45.6112s, MOSTLY ADDITIONS additions: 400400 removals: 20000
5.4808s, MOSTLY REMOVALS additions: 600 removals: 140000
57.1782s, FULLBLOCKS additions: 400400 removals: 400000
1.4021s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
0.9622s, GET RECORDS BY NAMES without spent 200 lookups found 20669 coins in total
19.2544s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 129.8889s`


`#SPINDLE original code
Building database
917.1889s, MOSTLY ADDITIONS additions: 400400 removals: 20000
325.2917s, MOSTLY REMOVALS additions: 600 removals: 140000
1281.8812s, FULLBLOCKS additions: 400400 removals: 400000
1.3874s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
0.8687s, GET RECORDS BY NAMES without spent 200 lookups found 20669 coins in total
19.5532s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 2546.1711s`

`#SPINDLE bulk insert and mass update
Building database
720.6202s, MOSTLY ADDITIONS additions: 400400 removals: 20000
305.5275s, MOSTLY REMOVALS additions: 600 removals: 140000
1087.8906s, FULLBLOCKS additions: 400400 removals: 400000
1.3928s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
0.7278s, GET RECORDS BY NAMES without spent 200 lookups found 20669 coins in total
19.3454s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 2135.5043s
`