@xorinox 2021-11-16
removed transaction id, benchmark still looks good
`Building database
40.1308s, MOSTLY ADDITIONS additions: 400400 removals: 20000
3.3671s, MOSTLY REMOVALS additions: 600 removals: 140000
49.1831s, FULLBLOCKS additions: 400400 removals: 400000
0.9352s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
1.3173s, GET RECORDS BY NAMES without spent 200 lookups found 20669 coins in total
19.1289s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 114.0625s`