@arvidn 2021-11-22
I pulled out your temporary table patch here: https://github.com/Chia-Network/chia-blockchain/pull/9355
I'm still running benchmarks for this one